backend/requirements.txt
openai>=1.0.0
langchain>=0.0.267
python-dotenv>=1.0.0
pyinstaller>=6.13.0
ollama>=0.1.0 